New Examples and AGeLib Update
- This commit includes the current AGeLib development snapshot (still WIP)
- Procedural terrain examples
- A hex map example (the strategy map example is still WIP as it lacks pathfinding and most of the map infrastructure)

Note: There are also some new files in the main folder that are intended for some future features. I only include these files in this commit because I am the only one committing to and using this repo. Otherwise some of the changes from this commit would need their own branches.