Qt compatibility
This commit makes a few minor adjustments to make APE fully compatible with PyQt6, PySide2, and PySide6. (PyQt5 is obviously still supported as before)

The `tobsprRenderPipeline` flag of the `start`-function is no longer overwritten (it was previously ignored (thus `start` always asked the user whether the pipeline should be used even if the application stated that it would not work) due to testing purposes and I forgot to implement a check for it...)

There are also some minor adjustments for the hex_grid_strategy_map example.