AGeLib Update: Minor Fixes and Notification Popup Experiment
Get rid of invalid escape sequence warnings.
Fixes type-hinting for App().Notification_List.
Fixes method and function execution in the Inspect widget of AGeIDE.
Notification button tool tip now eppears instantaneously and is also aligned to the notification button.
Adds code for an experimental NotificationPopup that might replace the notification button tool tip but that code is currently quite unstable and therefore deactivated. I like the concept but it might be more effort than it's worth and brings too much potential instability with it.